improved squareBadges
